chore: rename MobileBanner -> MobileTopNav
